# Various cache read and write performance optimizations.