--- a/src/cache/inmemory/writeToStore.ts
+++ b/src/cache/inmemory/writeToStore.ts
@@ -27,7 +27,7 @@ import { cloneDeep } from '../../utilities/common/cloneDeep';
 import { Policies } from './policies';
 import { defaultNormalizedCacheFactory } from './entityStore';
 import { NormalizedCache, StoreObject } from './types';
-import { makeProcessedFieldsMerger } from './helpers';
+import { makeProcessedFieldsMerger, FieldValueToBeMerged } from './helpers';
 
 export type WriteContext = {
   readonly store: NormalizedCache;
@@ -39,6 +39,16 @@ export type WriteContext = {
   // General-purpose deep-merge function for use during writes.
   merge<T>(existing: T, incoming: T): T;
 };
+
+interface ProcessSelectionSetOptions {
+  result: Record<string, any>;
+  selectionSet: SelectionSetNode;
+  context: WriteContext;
+  typename: string;
+  out: {
+    shouldApplyMerges?: boolean;
+  };
+}
 
 export interface StoreWriterConfig {
   policies: Policies;
@@ -134,20 +144,26 @@ export class StoreWriter {
       // fall back to that.
       store.get(dataId, "__typename") as string;
 
-    store.merge(
-      dataId,
-      policies.applyMerges(
+    const out: ProcessSelectionSetOptions["out"] = Object.create(null);
+
+    let processed = this.processSelectionSet({
+      result,
+      selectionSet,
+      context,
+      typename,
+      out,
+    });
+
+    if (out.shouldApplyMerges) {
+      processed = policies.applyMerges(
         makeReference(dataId),
-        this.processSelectionSet({
-          result,
-          selectionSet,
-          context,
-          typename,
-        }),
+        processed,
         store.getFieldValue,
         context.variables,
-      ),
-    );
+      );
+    }
+
+    store.merge(dataId, processed);
 
     return store;
   }
@@ -157,12 +173,10 @@ export class StoreWriter {
     selectionSet,
     context,
     typename,
-  }: {
-    result: Record<string, any>;
-    selectionSet: SelectionSetNode;
-    context: WriteContext;
-    typename: string;
-  }): StoreObject {
+    // This object allows processSelectionSet to report useful information
+    // to its callers without explicitly returning that information.
+    out,
+  }: ProcessSelectionSetOptions): StoreObject {
     let mergedFields: StoreObject = Object.create(null);
     if (typeof typename === "string") {
       mergedFields.__typename = typename;
@@ -186,22 +200,27 @@ export class StoreWriter {
             context.variables,
           );
 
-          const incomingValue =
-            this.processFieldValue(value, selection, context);
+          let incomingValue =
+            this.processFieldValue(value, selection, context, out);
 
-          mergedFields = context.merge(mergedFields, {
+          if (policies.hasMergeFunction(typename, selection.name.value)) {
             // If a custom merge function is defined for this field, store
             // a special FieldValueToBeMerged object, so that we can run
             // the merge function later, after all processSelectionSet
             // work is finished.
-            [storeFieldName]: policies.hasMergeFunction(
-              typename,
-              selection.name.value,
-            ) ? {
+            incomingValue = {
               __field: selection,
               __typename: typename,
               __value: incomingValue,
-            } : incomingValue,
+            } as FieldValueToBeMerged;
+
+            // Communicate to the caller that mergedFields contains at
+            // least one FieldValueToBeMerged.
+            out.shouldApplyMerges = true;
+          }
+
+          mergedFields = context.merge(mergedFields, {
+            [storeFieldName]: incomingValue,
           });
 
         } else if (
@@ -240,6 +259,7 @@ export class StoreWriter {
               selectionSet: fragment.selectionSet,
               context,
               typename,
+              out,
             }),
           );
         }
@@ -253,6 +273,7 @@ export class StoreWriter {
     value: any,
     field: FieldNode,
     context: WriteContext,
+    out: ProcessSelectionSetOptions["out"],
   ): StoreValue {
     if (!field.selectionSet || value === null) {
       // In development, we need to clone scalar values so that they can be
@@ -262,7 +283,8 @@ export class StoreWriter {
     }
 
     if (Array.isArray(value)) {
-      return value.map((item, i) => this.processFieldValue(item, field, context));
+      return value.map(
+        (item, i) => this.processFieldValue(item, field, context, out));
     }
 
     if (value) {
@@ -291,6 +313,7 @@ export class StoreWriter {
       context,
       typename: getTypenameFromResult(
         value, field.selectionSet, context.fragmentMap),
+      out,
     });
   }
 }