@benjamn 2020-02-17
These changes have been released in `@apollo/client@3.0.0-beta.35`, fyi.